Re-arranged the steps in a way that makes more sense
