+++ playwright.config.ts
@@ -56,7 +56,7 @@ export default defineConfig<TestOptions>({
                 outputDir: '.test-results/user',
                 importTestFrom: './fixtures/fixtureBuilder.ts',
                 paths: ['./features/setup/setup.feature'],
-                require: ['./steps/frontend/*.ts'],
+                require: ['./steps/*.ts'],
             }),
             use: { ...devices['Desktop Chrome'],
                 headless: process.env.USE_GUI != 'true',
@@ -70,7 +70,7 @@ export default defineConfig<TestOptions>({
                 outputDir: '.test-results/product',
                 importTestFrom: './fixtures/fixtureBuilder.ts',
                 paths: ['./features/backend/product.feature'],
-                require: ['./steps/backend/*.ts'],
+                require: ['./steps/*.ts'],
             }),
             dependencies: ['setup'],
         },
@@ -81,7 +81,7 @@ export default defineConfig<TestOptions>({
                 outputDir: '.test-results/frontend-chrome',
                 importTestFrom: './fixtures/fixtureBuilder.ts',
                 paths: ['./features/frontend/*.feature'],
-                require: ['./steps/frontend/*.ts'],
+                require: ['./steps/*.ts'],
             }),
             use: { ...devices['Desktop Chrome'],
                 headless: process.env.USE_GUI != 'true',
@@ -97,7 +97,7 @@ export default defineConfig<TestOptions>({
                 outputDir: '.test-results/safari-responsive',
                 importTestFrom: './fixtures/fixtureBuilder.ts',
                 paths: ['./features/frontend/*.feature'],
-                require: ['./steps/frontend/*.ts'],
+                require: ['./steps/*.ts'],
             }),
             use: { ...devices['iPhone 13'],
                 headless: process.env.USE_GUI != 'true',
@@ -113,7 +113,7 @@ export default defineConfig<TestOptions>({
                 outputDir: '.test-results/non-functional',
                 importTestFrom: './fixtures/fixtureBuilder.ts',
                 paths: ['./features/non-functional/*.feature'],
-                require: ['./steps/**/*.ts'],
+                require: ['./steps/*.ts'],
             }),
             use: { ...devices['Desktop Chrome'],
                 headless: process.env.USE_GUI != 'true',
