+++ steps/request.steps.ts
@@ -1,12 +1,13 @@
 import { expect } from '@playwright/test';
 import { createBdd } from 'playwright-bdd';
-import { test } from '../../fixtures/fixtureBuilder';
-import {CONFIG} from "../../variables.config";
-import {extractTestCaseTagID, getRandomItem} from "../../helpers/commonsHelper";
-import {Brand, Category, Product} from "../../helpers/typesHelper";
-import { productData } from "../../test-data/products.json" ;
+import { test } from '../fixtures/fixtureBuilder';
+import {CONFIG} from "../variables.config";
+import {extractTestCaseTagID, getRandomItem} from "../helpers/commonsHelper";
+import {Brand, Category, Product} from "../helpers/typesHelper";
+import { productData } from "../test-data/products.json" ;
+import {searchByNameStep} from "./productFiltering.steps"
 
-const { Given, Then } = createBdd(test);
+const { Given, When, Then } = createBdd(test);
 
 Given(/^I request a product to the Product API querying by brand$/, async ({ testContext, request, $tags }) => {
 
@@ -75,10 +76,28 @@ Given(/^I request a product to the Product API querying by category$/, async ({ 
 });
 
 
+When(/^I search for any product with the product API unavailable$/, async ({ filterPanel }) => {
+    await filterPanel.mockProductSearchAPIUnavailable();
+
+    const sampleProduct : Product = getRandomItem<Product>(productData);
+
+    await searchByNameStep({filterPanel}, sampleProduct.name);
+});
+
+
 Then(/^I find the requested product within the Product API response$/, async ({ testContext }) => {
-    const testProduct = testContext.getData('product');
-    const responseResults = (await (await testContext.retrieveResponse().json()).data).filter(d => d.name === testProduct.name);
+    const testProduct : Partial<Product> = testContext.getData('product');
+    const responseResults = (await (await testContext.retrieveResponse().json()).data).filter((d : Partial<Product>) => d.name === testProduct.name);
 
     await expect(responseResults[0]).toBeDefined();
     await expect(responseResults[0].name).toEqual(testProduct.name);
+});
+
+Then('I get a valid authorization token', async ({ request }) => {
+    //test authenticated request
+    expect((await request.get(CONFIG.baseApiHost + 'users/me',
+        {headers:
+                {'Authorization': `Bearer ${process.env.BEARER_TOKEN}`}
+        }))
+        .ok()).toBeTruthy();
 });
